// src/models/apiTaxComplianceTvrRecordListParams.ts
/**
 * Generated by orval
 * Do not edit manually.
 */

export type ApiTaxComplianceTvrRecordListParams = {
  created_at?: string
  created_at__gte?: string
  created_at__lte?: string
  filing_date?: string
  filing_date__gte?: string
  filing_date__lte?: string
  filing_frequency?: number
  /**
   * Multiple values may be separated by commas.
   */
  filing_frequency__in?: number[]
  filing_method?: number
  /**
   * Multiple values may be separated by commas.
   */
  filing_method__in?: number[]
  id?: number
  /**
   * Multiple values may be separated by commas.
   */
  id__in?: number[]
  is_ready?: boolean
  jurisdiction?: number
  /**
   * Multiple values may be separated by commas.
   */
  jurisdiction__in?: number[]
  legal_entity?: number
  /**
   * Multiple values may be separated by commas.
   */
  legal_entity__in?: number[]
  /**
   * Which field to use when ordering the results.
   */
  ordering?: string
  /**
   * A page number within the paginated result set.
   */
  page?: number
  /**
   * Number of results to return per page.
   */
  page_size?: number
  payment_conf_num?: string
  payment_conf_num__icontains?: string
  payment_date?: string
  payment_date__gte?: string
  payment_date__lte?: string
  period?: number
  /**
   * Multiple values may be separated by commas.
   */
  period__in?: number[]
  status_conf_num?: string
  status_conf_num__icontains?: string
  tax_type?: number
  /**
   * Multiple values may be separated by commas.
   */
  tax_type__in?: number[]
}